Moved calculate_gradient_along_node_links.
Moved calculate_gradient_along_node_links into raster_gradients.py
module.
